+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Improved
 - added additional Smartpost meta data adding when creating an order, if it failed to add the first time
 - added a ability to manually set Smartpost shipping method, when Smartpost data is missing in order
+- added a ability to hide shipping methods if cart weight or size is to big
 
 ## [1.4.0] - 2022-12-19
 ### Fixed